refactor(component): select field
- Create separate component for SelectItem to address linting issue
- Update SelectField to import separated SelectItem component
